requirement.txt
fastapi
uvicorn[standard]
elasticsearch
sentence-transformers
faiss-cpu